+++ src/config/config.go
@@ -1,9 +1,5 @@
 package config
 
-import (
-	"os/exec"
-)
-
 // Global system constants
 const NumButtons = 3
 const NumFloors = 4
@@ -24,14 +20,11 @@ const (
 // Local IP address
 var Laddr string
 
+// Global varibals
 var CurFloor int
 var CurDir int
+var ImConnected bool
 
-
-type Keypress struct {
-	Button int
-	Floor  int
-}
 
 // Generic network message. No other messages are ever sent on the network.
 type Message struct {
@@ -49,12 +42,6 @@ const (
 	CompleteOrder
 	Cost
 )
-
-var SyncLightsChan = make(chan bool)
-var CloseConnectionChan = make(chan bool)
-
-// Start a new terminal when restart.Run()
-var Restart = exec.Command("gnome-terminal", "-x", "sh", "-c", "main")
 
 // Colours for printing to console
 const Col0 = "\x1b[30;1m" // Dark grey
