+++ src/elevRun/elevFunc.go
@@ -29,8 +29,10 @@ func Elev_init(){
 
 func Run_elev(outgoingMsg chan def.Message) {
 	for {
-		update_current_floor()
+
 		floorSensor := driver.Get_floor_sensor_signal()
+		update_current_floor(floorSensor)
+
 		if at_top_or_bottom(floorSensor) {
 			def.CurDir = -def.CurDir
 		}
@@ -40,25 +42,45 @@ func Run_elev(outgoingMsg chan def.Message) {
 
 			if arrived_at_destination() {
 				//fmt.Printf("%sFloat cur floor = %v float orders[0] = %v %s \n", def.ColY, float64(floorSensor), math.Abs(float64(def.Orders[0])), def.ColN)
-				
-				send_complete_msg(queue.Get_Orders()[0],outgoingMsg)
-				//Removes the first element in def.Orders
+				driver.Set_motor_dir(0)
+
+				if def.ImConnected{
+					send_complete_msg(queue.Get_Orders()[0],outgoingMsg)
+				}else{
+					turn_off_lights()
+				}
+
+				remove_first_element_in_orders_and_save()
+
 				//fmt.Printf("%sOrder to floor %d deleted, current floor = %d\n",def.ColB,queue.Get_Orders()[0],def.CurFloor)
-				queue.Set_Orders(append(queue.Get_Orders()[:0], queue.Get_Orders()[1:]...))
-				queue.Save_backup_to_file()
+
 				driver.Set_button_lamp(def.BtnInside, def.CurFloor, 0)
 
-
-				driver.Set_motor_dir(0)
-				driver.Set_door_open_lamp(1)
-				time.Sleep(2 * time.Second)
-				driver.Set_door_open_lamp(0)
-				//Kjøre en func som setter motorDir her?
+				open_close_door()
 				
 
 			}
 		}
 	}
+}
+
+func turn_off_lights(){
+	if queue.Get_Orders()[0] < 0 {
+		driver.Set_button_lamp(def.BtnDown, def.CurFloor, 0)
+	} else {
+		driver.Set_button_lamp(def.BtnUp, def.CurFloor, 0)
+    }
+}
+
+func open_close_door(){
+	driver.Set_door_open_lamp(1)
+	time.Sleep(2 * time.Second)
+	driver.Set_door_open_lamp(0)
+}
+
+func remove_first_element_in_orders_and_save(){
+	queue.Set_Orders(append(queue.Get_Orders()[:0], queue.Get_Orders()[1:]...))
+	queue.Save_backup_to_file()
 }
 
 func set_direction(order,curFloor int){
@@ -90,57 +112,16 @@ func send_complete_msg(order int,outgoingMsg chan def.Message){
 	}
 }
 
-
-
-func Handle_orders(outgoingMsg chan def.Message) {
-	var alreadyPushed [def.NumFloors][def.NumButtons]bool
-
-	for {
-		for buttontype := 0; buttontype < 3; buttontype++ {
-			for floor := 0; floor < def.NumFloors; floor++ {
-				if button_pushed(buttontype,floor) {
-
-					if !alreadyPushed[floor][buttontype] {
-						driver.Set_button_lamp(buttontype, floor, 1)
-						handle_new_order(buttontype,floor,outgoingMsg)
-					}
-					alreadyPushed[floor][buttontype] = true
-				} else {
-					alreadyPushed[floor][buttontype] = false
-				}
-			}
-		}
-	}
-}
-
-func update_current_floor(){
-	var floorSensorSignal = driver.Get_floor_sensor_signal()
+func update_current_floor(floorSensorSignal int){
 	if floorSensorSignal != -1 {
 		def.CurFloor = floorSensorSignal
 		driver.Set_floor_indicator(floorSensorSignal)
 	}
 }
 
-func handle_new_order(buttontype,floor int,outgoingMsg chan def.Message){
-	if buttontype == def.BtnUp || buttontype == def.BtnDown {
-		msg := def.Message{Category: def.NewOrder, Floor: floor, Button: buttontype, Cost: -1}
-		outgoingMsg <- msg
-	} else {
-		//Internal orders: if the desired floor is under the elevator it is set as a order down
-		//fmt.Printf("Intern ordre mottat \n")
-		if floor < def.CurFloor {
-			queue.Set_Orders(queue.Update_orderlist(queue.Get_Orders(), -floor, false))
-		}else if floor == def.NumFloors-1 {
-			queue.Set_Orders(queue.Update_orderlist(queue.Get_Orders(), -floor, false))
-		}else{
-			queue.Set_Orders(queue.Update_orderlist(queue.Get_Orders(), floor, false))
-		}
 
-		queue.Save_backup_to_file()
-	}
-}
+/*--------------------------------------------------------------------------------------*/
 
-func button_pushed(buttontype,floor int)bool{
-	return driver.Get_button_signal(buttontype, floor) == 1
-}
+
+
