--- src/master/master.go
@@ -1,7 +1,0 @@
-package master
-
-func master_init(){
-/*	sjekke om det er andre på nettverket. Hvis ikke, så master
-	
-*/
-}
